requirements.txt
Django~=6.0
djangorestframework~=3.16.1
djangorestframework_simplejwt~=5.5.1
drf-yasg~=1.21.11
python-decouple~=3.8